git-quickinit/requirements.txt
PyGithub>=1.77